UI folder structure refactor ing
