+++ app/src/main/AndroidManifest.xml
@@ -28,7 +28,7 @@
             android:name=".MainActivity"
             android:screenOrientation="nosensor" />
         <activity
-            android:name=".ui.login.LoginActivity"
+            android:name=".LoginActivity"
             android:exported="true"
             android:screenOrientation="nosensor">
             <intent-filter>
