+++ app/src/main/java/ca/uwaterloo/treklogue/MainActivity.kt
@@ -16,11 +16,10 @@ import ca.uwaterloo.treklogue.data.repository.JournalEntryRealmSyncRepository
 import ca.uwaterloo.treklogue.data.repository.LandmarkRealmSyncRepository
 import ca.uwaterloo.treklogue.data.repository.UserRealmSyncRepository
 import ca.uwaterloo.treklogue.ui.Router
-import ca.uwaterloo.treklogue.ui.UserEvent
-import ca.uwaterloo.treklogue.ui.UserViewModel
-import ca.uwaterloo.treklogue.ui.login.LoginActivity
-import ca.uwaterloo.treklogue.ui.map.MapViewModel
-import ca.uwaterloo.treklogue.ui.profile.JournalEntryViewModel
+import ca.uwaterloo.treklogue.ui.viewModels.UserEvent
+import ca.uwaterloo.treklogue.ui.viewModels.UserViewModel
+import ca.uwaterloo.treklogue.ui.viewModels.MapViewModel
+import ca.uwaterloo.treklogue.ui.viewModels.JournalEntryViewModel
 import ca.uwaterloo.treklogue.ui.theme.MyApplicationTheme
 import kotlinx.coroutines.launch
