+++ app/src/main/java/ca/uwaterloo/treklogue/ui/composables/SettingsActionButton.kt
@@ -1,4 +1,4 @@
-package ca.uwaterloo.treklogue.ui.settings
+package ca.uwaterloo.treklogue.ui.composables
 
 import androidx.annotation.DrawableRes
 import androidx.annotation.StringRes
