+++ app/src/main/java/ca/uwaterloo/treklogue/ui/screens/ProfileScreen.kt
@@ -1,8 +1,10 @@
-package ca.uwaterloo.treklogue.ui.list
+package ca.uwaterloo.treklogue.ui.screens
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.rememberScrollState
 import androidx.compose.foundation.verticalScroll
@@ -12,28 +14,33 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.dp
 import ca.uwaterloo.treklogue.R
 import ca.uwaterloo.treklogue.data.mockModel.MockJournalEntry
-import ca.uwaterloo.treklogue.data.mockModel.MockLandmark
+import ca.uwaterloo.treklogue.ui.viewModels.UserViewModel
 import ca.uwaterloo.treklogue.ui.composables.TabSectionHeader
-import ca.uwaterloo.treklogue.ui.map.MapViewModel
-import ca.uwaterloo.treklogue.ui.profile.JournalEntryViewModel
+import ca.uwaterloo.treklogue.ui.composables.JournalEntryListItem
 import ca.uwaterloo.treklogue.ui.theme.Gray100
+import ca.uwaterloo.treklogue.ui.viewModels.JournalEntryViewModel
 
 /**
- * Composable for the list view
+ * Composable for the profile view
  */
 @Composable
-fun ListScreen(
+fun ProfileScreen(
     modifier: Modifier = Modifier,
-    mapViewModel: MapViewModel,
+    userViewModel: UserViewModel,
     journalEntryViewModel: JournalEntryViewModel,
-    onAddJournal: (landmark: MockLandmark) -> Unit
+    showJournalDetail: (journalEntry: MockJournalEntry) -> Unit,
 ) {
     Column(
         modifier = modifier.background(color = Gray100),
         horizontalAlignment = Alignment.Start,
         verticalArrangement = Arrangement.spacedBy(8.dp)
     ) {
-        TabSectionHeader(R.string.nearby_landmarks)
+        TabSectionHeader(R.string.my_badges)
+
+        // TODO: Badges (horizontally scrollable list?)
+        Spacer(modifier = Modifier.height(30.dp))
+
+        TabSectionHeader(R.string.my_journal_entries)
 
         Column(
             modifier = modifier
@@ -41,18 +48,15 @@ fun ListScreen(
                 .verticalScroll(rememberScrollState()),
             verticalArrangement = Arrangement.spacedBy(12.dp)
         ) {
-            mapViewModel.state.value.mockLandmarks.forEachIndexed { idx, landmark ->
-                LandmarkItem(
+            journalEntryViewModel.journalEntries.forEachIndexed { idx, journalEntry ->
+                JournalEntryListItem(
                     Modifier.padding(
                         top = if (idx == 0) 4.dp else 0.dp,
-                        bottom = if (idx == mapViewModel.state.value.landmarks.size - 1) 12.dp else 0.dp
+                        bottom = if (idx == journalEntryViewModel.journalEntries.size - 1) 12.dp else 0.dp
                     ),
-                    landmark,
-                    journalEntryViewModel,
-                    onAddJournal
+                    journalEntry, journalEntryViewModel, showJournalDetail
                 )
             }
         }
     }
-
 }
