+++ app/src/main/java/ca/uwaterloo/treklogue/ui/screens/SettingsScreen.kt
@@ -1,4 +1,4 @@
-package ca.uwaterloo.treklogue.ui.settings
+package ca.uwaterloo.treklogue.ui.screens
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Arrangement
@@ -11,9 +11,12 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.dp
 import ca.uwaterloo.treklogue.R
 import ca.uwaterloo.treklogue.app
-import ca.uwaterloo.treklogue.ui.UserEvent
-import ca.uwaterloo.treklogue.ui.UserViewModel
+import ca.uwaterloo.treklogue.ui.viewModels.UserEvent
+import ca.uwaterloo.treklogue.ui.viewModels.UserViewModel
 import ca.uwaterloo.treklogue.ui.composables.TabSectionHeader
+import ca.uwaterloo.treklogue.ui.composables.SettingsActionButton
+import ca.uwaterloo.treklogue.ui.composables.SettingsGroup
+import ca.uwaterloo.treklogue.ui.composables.SettingsToggle
 import ca.uwaterloo.treklogue.ui.theme.Gray100
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
